% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression_Description.R
\docType{package}
\name{Regression}
\alias{Regression}
\alias{Regression-package}
\title{Regression: A package for computating a linear regression model.}
\description{
The regression package includes two class of which one of then is for plotting.
}
\section{linreg function}{
calculates all the needed variables and saves then in an object called formulas,
}

