% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula, }{writen as a formula-object, see example.}

\item{data, }{a data.frame object in this particular case data(iris)}

\item{Lambda, }{numeric}
}
\value{
a ridgereg class-object containing regression outputs
}
\description{
This function calculates regression coeeficients and fitted values.
}
\details{
This function calculates Ridge regression
}
\examples{
ridreg(Sepal.Length ~ Sepal.Width, iris, 2)
}
\author{
Aqeel Ahmed, Marhawi Tewolde
}

